requirements_simple.txt
torch>=1.12.0
torchvision>=0.13.0
diffusers>=0.16.0
transformers>=4.26.0
numpy>=1.22.0
pillow>=9.0.0
tqdm>=4.64.0
trimesh>=3.16.0
opencv-python>=4.6.0
